requirements.txt
flask
flask-swagger-ui
pyserial
